\name{simData}
\docType{data}
\alias{simData}
\title{Simulated panel with some zeros}
\description{
This dataset is used to show how the nonlinear fit can be used when some
observations are zero. 
}
\usage{data(simData)}
\format{
A matrix containing 293 columns, and 5 time units.  
}
\keyword{datasets}
